+++ packages/contracts/contracts/registry/DAORegistry.sol
@@ -7,27 +7,27 @@ import {IDAO} from "../core/IDAO.sol";
 import {InterfaceBasedRegistry} from "./InterfaceBasedRegistry.sol";
 import {isSubdomainValid} from "./RegistryUtils.sol";
 
-/// @title Register your unique DAO name
+/// @title Register your unique DAO subdomain
 /// @author Aragon Association - 2022
 /// @notice This contract provides the possiblity to register a DAO.
 contract DAORegistry is InterfaceBasedRegistry {
     /// @notice The ID of the permission required to call the `register` function.
     bytes32 public constant REGISTER_DAO_PERMISSION_ID = keccak256("REGISTER_DAO_PERMISSION");
 
-    /// @notice The ENS subdomain registrar registering the DAO names.
+    /// @notice The ENS subdomain registrar registering the DAO subdomains.
     ENSSubdomainRegistrar private subdomainRegistrar;
 
-    /// @notice Thrown if the DAO repository name is empty.
-    error EmptyDaoName();
+    /// @notice Thrown if the DAO repository subdomain is empty.
+    error EmptyDaoSubdomain();
 
-    /// @notice Thrown if the DAO name doesn't match the regex `[0-9a-z\-]`
-    error InvalidDaoName(string name);
+    /// @notice Thrown if the DAO subdomain doesn't match the regex `[0-9a-z\-]`
+    error InvalidDaoSubdomain(string subdomain);
 
     /// @notice Emitted when a new DAO is registered.
     /// @param dao The address of the DAO contract.
     /// @param creator The address of the creator.
-    /// @param name The DAO name.
-    event DAORegistered(address indexed dao, address indexed creator, string name);
+    /// @param subdomain The DAO subdomain.
+    event DAORegistered(address indexed dao, address indexed creator, string subdomain);
 
     /// @dev Used to disallow initializing the implementation contract by an attacker for extra safety.
     constructor() {
@@ -46,32 +46,32 @@ contract DAORegistry is InterfaceBasedRegistry {
     }
 
     /// @notice Registers a DAO by its address.
-    /// @dev A name is unique within the Aragon DAO framework and can get stored here.
-    /// @param _dao The address of the DAO contract.
-    /// @param _creator The address of the creator.
-    /// @param _name The DAO name.
+    /// @dev A subdomain is unique within the Aragon DAO framework and can get stored here.
+    /// @param dao The address of the DAO contract.
+    /// @param creator The address of the creator.
+    /// @param subdomain The DAO subdomain.
     function register(
-        IDAO _dao,
-        address _creator,
-        string calldata _name
+        IDAO dao,
+        address creator,
+        string calldata subdomain
     ) external auth(REGISTER_DAO_PERMISSION_ID) {
-        address daoAddr = address(_dao);
+        address daoAddr = address(dao);
 
-        if (!(bytes(_name).length > 0)) {
-            revert EmptyDaoName();
+        if (!(bytes(subdomain).length > 0)) {
+            revert EmptyDaoSubdomain();
         }
 
-        if (!isSubdomainValid(_name)) {
-            revert InvalidDaoName({name: _name});
+        if (!isSubdomainValid(subdomain)) {
+            revert InvalidDaoSubdomain({subdomain: subdomain});
         }
 
         _register(daoAddr);
 
-        bytes32 labelhash = keccak256(bytes(_name));
+        bytes32 labelhash = keccak256(bytes(subdomain));
 
         subdomainRegistrar.registerSubnode(labelhash, daoAddr);
 
-        emit DAORegistered(daoAddr, _creator, _name);
+        emit DAORegistered(daoAddr, creator, subdomain);
     }
 
     /// @notice This empty reserved space is put in place to allow future versions to add new variables without shifting down storage in the inheritance chain (see [OpenZepplins guide about storage gaps](https://docs.openzeppelin.com/contracts/4.x/upgradeable#storage_gaps)).