+++ packages/contracts/test/registry/plugin-repo-registry.ts
@@ -11,7 +11,7 @@ import {
 import {deployNewDAO} from '../test-utils/dao';
 import {deployNewPluginRepo} from '../test-utils/repo';
 import {deployENSSubdomainRegistrar} from '../test-utils/ens';
-import {ensDomainHash, ensLabelHash} from '../../utils/ens';
+import {ensDomainHash} from '../../utils/ens';
 import {deployWithProxy} from '../test-utils/proxy';
 
 const EVENTS = {
@@ -35,7 +35,7 @@ describe('PluginRepoRegistry', function () {
   );
 
   const topLevelDomain = 'dao.eth';
-  const pluginRepoName = 'my-plugin-repo';
+  const pluginRepoSubdomain = 'my-plugin-repo';
 
   before(async () => {
     signers = await ethers.getSigners();
@@ -85,12 +85,12 @@ describe('PluginRepoRegistry', function () {
   it('Should register a new pluginRepo successfully', async function () {
     await expect(
       await pluginRepoRegistry.registerPluginRepo(
-        pluginRepoName,
+        pluginRepoSubdomain,
         pluginRepo.address
       )
     )
       .to.emit(pluginRepoRegistry, EVENTS.PluginRepoRegistered)
-      .withArgs(pluginRepoName, pluginRepo.address);
+      .withArgs(pluginRepoSubdomain, pluginRepo.address);
 
     expect(await pluginRepoRegistry.entries(pluginRepo.address)).to.equal(true);
   });
@@ -98,7 +98,7 @@ describe('PluginRepoRegistry', function () {
   it('fail to register if the sender lacks the required role', async () => {
     // Register a plugin successfully
     await pluginRepoRegistry.registerPluginRepo(
-      pluginRepoName,
+      pluginRepoSubdomain,
       pluginRepo.address
     );
 
@@ -114,7 +114,7 @@ describe('PluginRepoRegistry', function () {
 
     await expect(
       pluginRepoRegistry.registerPluginRepo(
-        pluginRepoName,
+        pluginRepoSubdomain,
         newPluginRepo.address
       )
     )
@@ -143,19 +143,19 @@ describe('PluginRepoRegistry', function () {
 
   it("reverts the registration if the plugin repo's ENS name is already taken", async function () {
     await pluginRepoRegistry.registerPluginRepo(
-      pluginRepoName,
+      pluginRepoSubdomain,
       pluginRepo.address
     );
 
-    const pluginRepoNameDomainHash = ensDomainHash(
-      pluginRepoName + '.' + topLevelDomain
+    const pluginRepoSubdomainDomainHash = ensDomainHash(
+      pluginRepoSubdomain + '.' + topLevelDomain
     );
 
     await expect(
-      pluginRepoRegistry.registerPluginRepo(pluginRepoName, pluginRepo.address)
+      pluginRepoRegistry.registerPluginRepo(pluginRepoSubdomain, pluginRepo.address)
     )
       .to.be.revertedWithCustomError(ensSubdomainRegistrar, 'AlreadyRegistered')
-      .withArgs(pluginRepoNameDomainHash, ensSubdomainRegistrar.address);
+      .withArgs(pluginRepoSubdomainDomainHash, ensSubdomainRegistrar.address);
   });
 
   // without mocking we have to repeat the tests here to make sure the validation is correct
@@ -193,7 +193,7 @@ describe('PluginRepoRegistry', function () {
         )
           .to.be.revertedWithCustomError(
             pluginRepoRegistry,
-            'InvalidPluginName'
+            'InvalidPluginSubdomain'
           )
           .withArgs(subdomainName);
       }
@@ -233,7 +233,7 @@ describe('PluginRepoRegistry', function () {
         )
           .to.be.revertedWithCustomError(
             pluginRepoRegistry,
-            'InvalidPluginName'
+            'InvalidPluginSubdomain'
           )
           .withArgs(subdomainName);
       }